Refactored Upload Profile Picture to adhere to MVC file structure.
