+++ components/profile/profilecomponents/profilePicture/UploadProfilePictureController.js
@@ -1,14 +1,17 @@
 // import React in our code
 import React, { useEffect, useState } from "react";
 // import all the components we are going to use
-import { StyleSheet, View, Image, Alert, Platform } from "react-native";
-
-import { Text, Button } from "react-native-elements";
+import { Alert, Platform } from "react-native";
 import * as ImagePicker from "expo-image-picker";
 import { storage } from "../../../../firebase/firebaseConfig";
 import firestoreService from "../../../../firebase/firestoreService";
+import UploadProfilePictureView from "./UploadProfilePictureView";
 
-function UploadProfilePicture({ setProfilePicture, toggleOverlay, user }) {
+function UploadProfilePictureController({
+  setProfilePicture,
+  toggleOverlay,
+  user,
+}) {
   const [image, setImage] = useState("");
 
   // Only run on android and ios. Webpack does not have Logbox.
@@ -110,68 +113,15 @@ function UploadProfilePicture({ setProfilePicture, toggleOverlay, user }) {
     })();
   }, []);
 
-  const renderPage = () => {
-    return (
-      <View style={styles.container}>
-        <Text h3 style={{ textAlign: "center", marginBottom: 20 }}>
-          Upload Profile Picture
-        </Text>
-        <Button
-          title="Choose Photo"
-          onPress={imagePicker}
-          buttonStyle={styles.globalButton}
-        />
-        {image ? (
-          <Image style={styles.image} source={{ uri: image }} />
-        ) : (
-          <Image
-            style={styles.image}
-            source={require("../../../../assets/icon.png")}
-          ></Image>
-        )}
-        <Button
-          title="Upload Profile Picture"
-          onPress={upload}
-          buttonStyle={styles.globalButton}
-        />
-        <Button
-          title="Remove Profile Picture"
-          onPress={removePictureAlert}
-          buttonStyle={[styles.globalButton, styles.removeButton]}
-        />
-        <Button
-          title="Cancel"
-          onPress={toggleOverlay}
-          buttonStyle={styles.globalButton}
-        />
-      </View>
-    );
-  };
-
-  return renderPage();
+  return (
+    <UploadProfilePictureView
+      imagePicker={imagePicker}
+      image={image}
+      upload={upload}
+      removePictureAlert={removePictureAlert}
+      toggleOverlay={toggleOverlay}
+    />
+  );
 }
 
-export default UploadProfilePicture;
-
-const styles = StyleSheet.create({
-  container: {
-    alignItems: "center",
-    padding: 15,
-  },
-  image: {
-    width: 200,
-    height: 200,
-    marginBottom: 20,
-    marginTop: 20,
-    borderRadius: 100,
-    borderColor: "black",
-    borderWidth: 2,
-  },
-  globalButton: {
-    borderRadius: 10,
-    marginTop: 20,
-  },
-  removeButton: {
-    backgroundColor: "red",
-  },
-});
+export default UploadProfilePictureController;
